Refactored flights
