+++ operator-service/Tarom-service/src/main/java/org/example/flight/dto/FlightDto.java
@@ -1,7 +1,9 @@
 package org.example.flight.dto;
 
 import jakarta.validation.constraints.NotNull;
-import org.example.flight_details.dto.FlightDetailsDto;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 
 public class FlightDto {
 
@@ -12,6 +14,51 @@ public class FlightDto {
     private String departure;
     @NotNull(message = "Destination cannot be null")
     private String destination;
+    private LocalDate date;
+    private LocalDateTime departureTime;
+    private LocalDateTime arrivalTime;
+    private int numberOfSeats;
+    private double standardPrice;
+
+    public LocalDate getDate() {
+        return date;
+    }
+
+    public void setDate(LocalDate date) {
+        this.date = date;
+    }
+
+    public LocalDateTime getDepartureTime() {
+        return departureTime;
+    }
+
+    public void setDepartureTime(LocalDateTime departureTime) {
+        this.departureTime = departureTime;
+    }
+
+    public LocalDateTime getArrivalTime() {
+        return arrivalTime;
+    }
+
+    public void setArrivalTime(LocalDateTime arrivalTime) {
+        this.arrivalTime = arrivalTime;
+    }
+
+    public int getNumberOfSeats() {
+        return numberOfSeats;
+    }
+
+    public void setNumberOfSeats(int numberOfSeats) {
+        this.numberOfSeats = numberOfSeats;
+    }
+
+    public double getStandardPrice() {
+        return standardPrice;
+    }
+
+    public void setStandardPrice(double standardPrice) {
+        this.standardPrice = standardPrice;
+    }
 
     public String getId() {
         return id;
