+++ operator-service/Tarom-service/src/main/java/org/example/flight/model/Flight.java
@@ -3,6 +3,9 @@ package org.example.flight.model;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+
 @Document(collection = "flights")
 public class Flight {
     @Id
@@ -11,6 +14,51 @@ public class Flight {
     private String operatorId;
     private String departure;
     private String destination;
+    private LocalDate date;
+    private LocalDateTime departureTime;
+    private LocalDateTime arrivalTime;
+    private int numberOfSeats;
+    private double standardPrice;
+
+    public LocalDate getDate() {
+        return date;
+    }
+
+    public void setDate(LocalDate date) {
+        this.date = date;
+    }
+
+    public LocalDateTime getDepartureTime() {
+        return departureTime;
+    }
+
+    public void setDepartureTime(LocalDateTime departureTime) {
+        this.departureTime = departureTime;
+    }
+
+    public LocalDateTime getArrivalTime() {
+        return arrivalTime;
+    }
+
+    public void setArrivalTime(LocalDateTime arrivalTime) {
+        this.arrivalTime = arrivalTime;
+    }
+
+    public int getNumberOfSeats() {
+        return numberOfSeats;
+    }
+
+    public void setNumberOfSeats(int numberOfSeats) {
+        this.numberOfSeats = numberOfSeats;
+    }
+
+    public double getStandardPrice() {
+        return standardPrice;
+    }
+
+    public void setStandardPrice(double standardPrice) {
+        this.standardPrice = standardPrice;
+    }
 
     public String getId() {
         return id;
