+++ operator-service/Tarom-service/src/main/java/org/example/flight/repository/FlightRepository.java
@@ -5,6 +5,8 @@ import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.time.LocalDate;
+
 public interface FlightRepository extends ReactiveMongoRepository<Flight,String> {
-    Flux<Flight> findByDepartureAndDestination(String departure, String destination);
+    Flux<Flight> findByDepartureAndDestinationAndDate(String departure, String destination, LocalDate date);
 }
