--- operator-service/Tarom-service/src/main/java/org/example/flight_details/dto/FlightDetailsDto.java
@@ -1,70 +1,0 @@
-package org.example.flight_details.dto;
-
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-
-public class FlightDetailsDto {
-    private String id;
-    private String flightId;
-    private LocalDate date;
-    private LocalDateTime departureTime;
-    private LocalDateTime arrivalTime;
-    private int numberOfSeats;
-    private double standardPrice;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getFlightId() {
-        return flightId;
-    }
-
-    public void setFlightId(String flightId) {
-        this.flightId = flightId;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public void setDate(LocalDate date) {
-        this.date = date;
-    }
-
-    public LocalDateTime getDepartureTime() {
-        return departureTime;
-    }
-
-    public void setDepartureTime(LocalDateTime departureTime) {
-        this.departureTime = departureTime;
-    }
-
-    public LocalDateTime getArrivalTime() {
-        return arrivalTime;
-    }
-
-    public void setArrivalTime(LocalDateTime arrivalTime) {
-        this.arrivalTime = arrivalTime;
-    }
-
-    public int getNumberOfSeats() {
-        return numberOfSeats;
-    }
-
-    public void setNumberOfSeats(int numberOfSeats) {
-        this.numberOfSeats = numberOfSeats;
-    }
-
-    public double getStandardPrice() {
-        return standardPrice;
-    }
-
-    public void setStandardPrice(double standardPrice) {
-        this.standardPrice = standardPrice;
-    }
-}
