--- operator-service/Tarom-service/src/main/java/org/example/flight_details/exception/FlightDetailsExceptionHandler.java
@@ -1,22 +1,0 @@
-package org.example.flight_details.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.time.LocalDateTime;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-@ControllerAdvice
-public class FlightDetailsExceptionHandler {
-
-    @ExceptionHandler(FlightDetailsNotFoundException.class)
-    public ResponseEntity<Object> handleFlightDetailsNotFoundException(FlightDetailsNotFoundException ex) {
-        Map<String, Object> body = new LinkedHashMap<>();
-        body.put("timestamp", LocalDateTime.now());
-        body.put("message", ex.getMessage());
-        return new ResponseEntity<>(body, HttpStatus.NOT_FOUND);
-    }
-}
