--- operator-service/Tarom-service/src/main/java/org/example/flight_details/exception/FlightDetailsNotFoundException.java
@@ -1,7 +1,0 @@
-package org.example.flight_details.exception;
-
-public class FlightDetailsNotFoundException extends RuntimeException {
-    public FlightDetailsNotFoundException(String id) {
-        super("Flight details with id: " + id + " not found!");
-    }
-}
